♻️ Refacto fetch users notifications
